pyproject.toml
[project]
name = "api-embrapa-vitivinicultura"
version = "2025.05.001"
description = ""
authors = [
    {name = "Isaque Vacari",email = "isaque.vacari@embrapa.br"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flasgger (>=0.9.7.1,<0.10.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "cryptography (>=44.0.3,<45.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "flask-caching (>=2.3.1,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"

